Permissions testing

+++ app/Http/Controllers/ShiftController.php
@@ -21,7 +21,10 @@ class ShiftController extends Controller
     public function store(Request $request)
     {
         $data = $this->storeValidator($request->all())->validate();
-        $this->checkPermissions($request);
+
+        if ($this->userHasNoPermissions($request)) {
+            return response()->json(['error' => 'Access denied.'], ResponseStatus::HTTP_FORBIDDEN);
+        }
 
         $data['created_by'] = Auth::id();
 
@@ -37,14 +40,17 @@ class ShiftController extends Controller
     /**
      * Update existing data
      *
-     * @param Response $response
+     * @param Request $request
      * @param Shift $shift
      * @return json
      */
-    public function update(Request $request, Shift $shift)
+    public function update(Shift $shift, Request $request)
     {
         $data = $this->updateValidator($request->all())->validate();
-        $this->checkPermissions($request);
+
+        if ($this->userHasNoPermissions($shift)) {
+            return response()->json(['error' => 'Access denied.'], ResponseStatus::HTTP_FORBIDDEN);
+        }
         $data['updated_by'] = Auth::id();
 
         if (!$shift->update($data)) {
@@ -62,7 +68,10 @@ class ShiftController extends Controller
      */
     public function destroy(Shift $shift)
     {
-        $this->checkPermissions($shift);
+        if ($this->userHasNoPermissions($shift)) {
+            return response()->json(['error' => 'Access denied.'], ResponseStatus::HTTP_FORBIDDEN);
+        }
+
         if (!$shift->delete()) {
             return response()->json(['error' => 'An error occured.'], ResponseStatus::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -98,8 +107,6 @@ class ShiftController extends Controller
     protected function updateValidator(array $data)
     {
         return Validator::make($data, [
-            'worker_id' => 'numeric|min:0',
-            'work_place_id' => 'required|numeric|min:0',
             'day' => 'date_format:Y-m-d',
             'shift_start' => 'date_format:H:i',
             'shift_end' => 'date_format:H:i',
@@ -112,16 +119,15 @@ class ShiftController extends Controller
      * @param Request $request
      * @return json : void
      */
-    protected function checkPermissions($item)
+    protected function userHasNoPermissions($item)
     {
         if ($item instanceof Request) {
+            $workPlace1 = $item->get('work_place_id');
             $workPlace = WorkPlace::findOrFail($item->get('work_place_id'));
         } elseif ($item instanceof Shift) {
             $workPlace = $item->workPlace;
         }
 
-        if (Gate::denies('edit', $workPlace)) {
-            return response()->json(['error' => 'Access denied.'], ResponseStatus::HTTP_FORBIDDEN);
-        }
+        return (Gate::denies('edit', $workPlace));
     }
 }
